// src/types.ts
export enum SlideLayout {
  WITHOUT = "without",
  TOP_IMAGE = "top-image",
  LEFT_IMAGE = "left-image",
  RIGHT_IMAGE = "right-image",
  TITUL = "titul",
  SECTION_INTRO = "section-intro",
  FINAL = "final",
}

export enum SlideVerticalAlign {
  TOP = "top",
  CENTER = "center",
  BOTTOM = "bottom",
}

export type Slide = {
  id: string;
  title?: string;
  content: string;
  verticalAlign?: SlideVerticalAlign;
  colorAccent?: "light" | "dark";
  backgroundImageUrl?: string;
  layout?: SlideLayout;
  layoutImageUrl?: string;
  isLoadingLayoutImage?: boolean;
  index?: number;
};